add AuthController & Middleware

+++ database/migrations/2026_05_18_124016_add_role_to_users_table.php
@@ -13,17 +13,21 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::table('users', function (Blueprint $table) {
-            $table->enum('role', ['mahasiswa', 'admin', 'guest'])
-                  ->default('mahasiswa')
-                  ->after('email');
-        });
+        if (!Schema::hasColumn('users', 'role')) {
+            Schema::table('users', function (Blueprint $table) {
+                $table->enum('role', ['mahasiswa', 'admin', 'guest'])
+                      ->default('mahasiswa')
+                      ->after('email');
+            });
+        }
     }
 
     public function down(): void
     {
-        Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn('role');
-        });
+        if (Schema::hasColumn('users', 'role')) {
+            Schema::table('users', function (Blueprint $table) {
+                $table->dropColumn('role');
+            });
+        }
     }
 };
